Fix random links
